- fixing some problems in linux build (typename issues) - OPEN ISSUE: aabp_tree suddenly broke with an ASSERT FAILURE
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@386 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/io/image.cpp
@@ -711,13 +711,13 @@ BinaryIStream& Image::openTargaTrueColor(BinaryIStream& iStream, const HeaderTar
 	};
 
 	LASS_ASSERT(cols_ == iHeader.imageWidth);
-	const int xBegin = iHeader.flipHorizontalFlag() ? cols_ - 1 : 0;
-	const int xEnd   = iHeader.flipHorizontalFlag() ? -1 : cols_;
+	const int xBegin = iHeader.flipHorizontalFlag() ? static_cast<int>(cols_) - 1 : 0;
+	const int xEnd   = iHeader.flipHorizontalFlag() ? -1 : static_cast<int>(cols_);
 	const int xDelta = iHeader.flipHorizontalFlag() ? -1 : 1;
 
 	LASS_ASSERT(rows_ == iHeader.imageHeight);
-	const int yBegin = iHeader.flipVerticalFlag() ? 0 : rows_ - 1;
-	const int yEnd   = iHeader.flipVerticalFlag() ? rows_ : -1;
+	const int yBegin = iHeader.flipVerticalFlag() ? 0 : static_cast<int>(rows_) - 1;
+	const int yEnd   = iHeader.flipVerticalFlag() ? static_cast<int>(rows_) : -1;
 	const int yDelta = iHeader.flipVerticalFlag() ? 1 : -1;
 
 	iStream.seekg(iHeader.idLength + iHeader.colorMapLength * iHeader.colorMapEntrySize,
